pyproject.toml
[build-system]
requires = ["setuptools>=77.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "koreanstocks"
version = "0.2.3"
description = "KOSPI·KOSDAQ 종목 자동 스크리닝 + AI/ML 분석 + 텔레그램 리포트 플랫폼"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
license-files = ["LICENSE"]
authors = [
    {name = "Sungwoo Kim", email = "sungwoo.kim@gmail.com"}
]
keywords = [
    "korean-stocks", "kospi", "kosdaq", "stock-screening",
    "technical-analysis", "machine-learning", "sentiment-analysis",
    "fastapi", "telegram", "finance", "investment",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "Intended Audience :: Developers",
    "Topic :: Office/Business :: Financial :: Investment",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
]
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "typer>=0.12",
    "openai>=1.10,<2",
    "pandas>=2.2,<4",
    "scikit-learn>=1.6,<2",
    "xgboost>=2,<4",
    "pykrx>=1.0.45",
    "finance-datareader>=0.9.50",
    "ta>=0.11",
    "plotly>=5.18",
    "python-dotenv>=1,<2",
    "requests>=2.31",
    "numpy>=2.0,<3",
    "joblib>=1.3,<2",
    "scipy>=1.11,<2",
    "beautifulsoup4>=4.12",
    "schedule>=1.2,<2",
    "matplotlib>=3.8",
    "httpx>=0.24,<1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23",
    "black>=23.7.0",
    "flake8>=6.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.3.0",
]

[project.urls]
Homepage = "https://github.com/bullpeng72/KoreanStock"
Repository = "https://github.com/bullpeng72/KoreanStock"
"Bug Tracker" = "https://github.com/bullpeng72/KoreanStock/issues"

[project.scripts]
koreanstocks = "koreanstocks.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"koreanstocks" = ["static/**/*"]

[tool.black]
line-length = 120
target-version = ["py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | venv
  | build
  | dist
  | \.egg-info
  | data
  | models
)/
'''

[tool.isort]
profile = "black"
line_length = 120
skip = [".venv", "venv", "build", "dist", "data", "models"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = false
disallow_untyped_defs = false
check_untyped_defs = true
